virsh_migrate_setmaxdowntime: Fix conf issue in modular daemon mode
Translation of conf_type from monolithic daemon mode to modular daemon
mode is complex and depends on the parameter to be configured heavily,
while the reverse translation is much more simple by tranlating all the
conf_type of modular daemon mode to "libvirtd". This commit sets the
conf_type to the value of modular daemon mode in <test>.cfg and the
<test>.py or avocado-vt utils will translate to "libvirtd" if test env
is monolithic daemon mode.

Signed-off-by: Fangge Jin <fjin@redhat.com>