[nodedev-dumpxml] add dasd support on s390x
The dumpxml information is used to set up a vfio_ccw passthrough device.
The device hierarchy by cap is dasd -> ccw -> css. The css id is used to
configure the mdev.

Test minimally that the chain is valid and that the device are
of the right type for dasd passthrough.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>