[migrate_vm] enable kvm hpage via configuration
Some hosts might load kvm without hugepage support, e.g. on s390x
because hugepage support is currently incompatible with nested support.

Add logic to test cases `with_hugepages` so that kvm is reloaded with
hugepage support on source and/or target host, if configured.

Default is "no", that is the code will work as before.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>