controller_functional: fix vm define error
This is to fix a vm define error like below. This error happened due
to a new libvirt parsing behavior in 7.0.0-x. But the fix is compatible
to old libvirt versions.

XML error: The PCI controller with index='0' must be model='pcie-root'
for this machine type, but model='pci-bridge' was found instead

Signed-off-by: Dan Zheng <dzheng@redhat.com>