[virsh_blockjob] Introduce configuration to slow down blockjob
1. The tests start a blockjob (blockcopy) and then execute
several blockjob commands often expecting the blockjob to be
running.

2. The test waits 3 seconds after starting the blockjob.

As a result, if the blockjob ends before 3 seconds have passed
there will be no blockjob and the domain will have been stopped.
Therefore, the test will fail.

Introduce the bandwidth parameter value. It will cause the blockjob
to take more time. Value 300 was chosen to not slow down the .with_pivot
subcase too much as it waits for the job to finish.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>