Make the cfg file more robust
As we will update the first interface to be the one exists on the host,
and do not care the others(the others will not be used by the guest, so
it is acceptable), if we test duplicate dev, we should ensure the others
are duplicate. For net_private test, no need 2 deivces as it may cause
error on certain systems.

Signed-off-by: Yalan <yalzhang@redhat.com>