v2v: refactor the video and qxl driver checkpoints
In different v2v and rhel versions, the video and qxl driver may
be set to different values, and since the qxl are going to be
removed on rhel9, the old code won't adapt those changes and is
difficult to maintain. This patch makes a change to ease the pain.

1) Move the logic of detecting virtio-win to a single function, it
can be reused by other functions.
2) Separate the expected video logic per different target, and Modifies
the logic of expected video.
3) Fixes the expected qxl driver error in windows guests when video is
not 'qxl'.
4) Give one more try when exception happens during the checking of windows guest.

Signed-off-by: Xiaodai Wang <xiaodwan@redhat.com>