Add more model to the domif-setlink test
The set link up and down are related with the interface driver.
So expand the test coverage to test all current supported models.
Add add 1 check point for several cases to check the link state
in vm. As the error_msg may be override, raise it once error_msg
is not None, it is helpful for debugging.

Signed-off-by: Yalan <yalzhang@redhat.com>