virsh_event: relax regex
7494d48d91b9e7d75a8b76b3ef9a79b05f4c796e changed the regex
to match error messages against to reflect newer versions
that use "'". However, older versions now have a test failure.
Instead of introducing a switch over the actual libvirt
version have a less strict regex to confirm that the domain
name is part of the error message independently from the usage
of "'".

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>