migrate_storage: improve test when migrate again
8fd5e336fc03936594927fe89ab0d8deaac5cd2e introduced
`--bandwidth 10` to make sure migration can be cancelled
in the cancel-retry test cases. However, this slows the
actual test run down and can lead to timeouts resulting
in a failed test case.

Introduce an additional parameter for all those test cases
that run migration again. If not given, it will default to the
given extra parameters in order to maintain default behavior.
However, the parameter is set in the cancel test case in
order to avoid timeouts by using default bandwidth (not passing
the `--bandwith 10` flag again).

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>