Fix the guest_agent guestinfo case for hostname
In the past, there were no hostname command available in some images
and therefore the hostnamectl was used instead. The hostname is available
again and the hostnamectl causing issues right now and therefore the
hostname command was added back, but the hostnamectl is there as a backup
for images, where the hostname might not be available.

Signed-off-by: Kamil Varga <kvarga@redhat.com>