virsh_detach_device_alias: fix detach slow issue
The cmd "detach-device-alias" will return "Device detach request sent successfully" immediately,
which may cause that the related device info will not take effect in the dumpxml/guest os that fast.
Simply waiting for few seconds can not make the script stable on different test environments. This
is to enhance it.

Signed-off-by: Dan Zheng <dzheng@redhat.com>