virsh_guestinfo: Fix filesystem test error caused by the wrong order
Filesystem test compares the output of 'virsh guestinfo --filesystem' with
guest lsblk cmd ouput.

This test sorted the output of lsblk cmd in the past to keep pace with
agent output.
But the order of agent is random in fact. When guest has more than 2 fs,
test is easy to be ERROR due to unmatching info.

Signed-off-by: Liu Yiding <liuyd.fnst@fujitsu.com>