migrate_vm: simplify target hugepage size formula
Commit e6ff5b9c6d277696c884a3869ba38b5f22a5123b updated the formula
that calculates the needed target hugepages for successful and unsuccessful
migration with hugepages on the destination host.
This way it fixed failing negative test cases (unsuccessful migration).
However, now the positive test case fails on a system with
Guest VM memory MB: 1024
Hugepagesize: 1024 KB

The formula seems to assume a specific hugepage size (1024*2) and looks rather
complicated.

It is difficult to determine this number as it depends on the number
of already allocated hugepages locally and on the different hugepage sizes.
For the source host this is handled through the 'setup_hugepages = yes' parameter
in the test configuration that is used during 'env_process.py' to set up
a correct hugepage number so that vms can run successfully with memory backed
by hugepages.

If we assume a specific setup (s. comment in code) we can use a simpler
formula and still have both test scenarios (negative, postive) pass.

Additionally, log some global parameters that have an impact on how
'env_process.py' sets up the hugepages on the source host.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>