libvirt_rng: relax test condition inside guest
The test case currently expects virtio to be rng hw_device
that's selected by the kernel but this depends on the
kernel code. Let's just check that some hw_device is set
and the virtio device is available. The way the kernel selects
the device is unlikely to change unintentionally and
the current_rng can always be set manually.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>