vcpu_nested: adapt for modular daemons
1 - Checking log is sometimes not statble. After confirming with feature
owner, this step is not required for this case. So I remove it to
make the script stabler.

2 - After libvirt packages are installed in L1, to make the default
libvirt daemons start automatically, we have 2 ways. One is rebooting
the vm, the other is to explicitly start the service names according
to different libvirt versions to judge monolithic vs. modular daemons
to use. It seems the first way is easier.

Signed-off-by: Dan Zheng <dzheng@redhat.com>