watchdog: relax qemu cmdline check
The used qemu interface for watchdog devices has changed from
version 6.1 to 6.2.

Compare

6.1: -device diag288,id=watchdog0 -watchdog-action inject-nmi
6.2: -device {"driver":"diag288","id":"watchdog0"} -watchdog-action
  inject-nmi

Therefore, tests failed with qemu 6.2 because they checked the qemu
cmdline according to 6.1 interface.

Update the test code to only check if the name is listed. This should
sufficient, esp. considering that the actual actions are tested later
on. The variable 'watchdog_device' is only used for the cmdline check.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>