v2v: fix qxl check for win11 and win2022 guests
For rhel9, we don't need to do any qxl check because it's not supported.

For rhel8, because 1997446 and 2012658, the win11 and win2022 are
inspected to win10 and win2016 unexpectedly by libguestfs, and this is
not going to be fixed in rhel8. so the qxl driver is installed for win11
and win2022 by mistake. BTW, virtio-win has already removed qxldod
driver for win11 and win2022. Because this problem, in order to make
out test run successfully, we'll continue checking the qxl for win11 and
win2022, but for the qxl warning message in v2v debug log, the scrip
will skip checking it. This warning is not that important and will make
the logic more complicate to understand.

Signed-off-by: Xiaodai Wang <xiaodwan@redhat.com>