nodedev_persistence: add test case for nodedev persistent setup API
Add a roundtrip test case to cover the lifecycle of a ccw mdev device
that's set up persistently via libvirt's nodedev API.

As avocado can't restart the host and continue the test run,
the autostart configuration is only covered to be correctly set.

All other commands are expected to fail if there was any issue.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>