v2v: update the dependency checking command
'yum deplist xxx' sometimes cannot query out correct result because
the db is not latest. I'm not sure how it happens. e.g. if you install
the pkg by a url link instead of repos, it always query from the old
package. so I think 'rpm -qR xxx' is better than that in this case.

Signed-off-by: Xiaodai Wang <xiaodwan@redhat.com>