Fix variable referenced before assignment error
This PR fixed the error:
UnboundLocalError: local variable 'config' referenced before assignment.
The path for config file no longer exists in some new RHEL versions.
which cause variable "config" initialize failed before used,
and then the error UnboundLocalError occured.

To avoid this failure,we decide to delete the code about sysconfig
in src and cfg as the sysconfig is not a test point,deleting it will not
affect the case and test can pass in any circumstance without it.

Signed-off-by: haonanpan <hpan@redhat.com>