virsh_domblkerror: fix mkfs.ext3 timeout
There is a conflict mkfs command for iscsi disk which causes the cases unstable.
For example, the newly created iscsi disk is executed with 'mkfs.ext4 -F /dev/sdb ',
then this disk passed to vm and is executed with 'mkfs.ext3 /dev/vdb1' which makes
conflict. So sometimes the case will fail with
'Timeout expired while waiting for shell command to complete: 'mkfs.ext3 /dev/vdb1''