provider/vfio/ccw: improve logging
I suspect fdasd might succeed sometimes despite a non-zero
status code though I couldn't reproduce it. This made the test
case fail. (I think this could happen because of a warning
when the disk is detected as used by several hosts.)

Add the status report to the log in case this reproduces.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>