Add -device to list of args with redundant quotes
In virsh_domxml_to_native.py, when exporting setup from xml,
for some arguments quotes are added. The testing script removes them but
the list of arguments that is supposed to be removed is edited manually.

One of the missing arguments is -device arguments that this commit adds.

Similar commit is bab218c2074e3537db1b18e0d0c8e1c4bbc9cd04
https://github.com/autotest/tp-libvirt/commit/bab218c2074e3537db1b18e0d0c8e1c4bbc9cd04