Fix the issues about flock/posix
1) As shown in https://bugzilla.redhat.com/show_bug.cgi?id=2057882#c9,
posix/flock is not supported, some import scenarios in only tested when
posix/flock=on before, and these scenarios have be added into skiplist.
We should change posix/flock to off in these scenarios.

2) If I run the cases on my local machine, there would be a process of
avocado contains "virtiofsd". If one just grep the first process of
virtiofsd, it is not the virtiofsd process itself.

Signed-off-by: Lili Zhu <lizhu@redhat.com>