v2v: add a variant to control the vsock checking
Whether a VM suppots vsock or not depends on the vm's kernel, it's
difficult in our script to check it and not necessary. For most of
cases, the vsock check can be ignored except the matrix and acceptance
and signoff testing.

So we can enabled the check in convert_vm_to_libvirt but other files
to avoid unexpected vsock checking failure.

Signed-off-by: Xiaodai Wang <xiaodwan@redhat.com>