Update scsi controller on ppc vms to support scsi hotplug
The code deletes the scsi controller but not replacing with "virtio-scsi" due
to which VM boots fails.

Reference - commit 1473a1a

The comment "Remove old scsi controller, replace with virtio-scsi controller"
is inaccurate. The code is removing the old controller, but it is not
replacing it with anything. It seems to be assuming that libvirt will default
to adding a virtio-scsi controller, but that is not the case. libvirt defaults
to the spapr-vscsi controller on ppc64le, not virtio-scsi. So
update_controllers_ppc should instead explicitly add a virtio-scsi controller.

Signed-off-by: Seeteena Thoufeek s1seetee@linux.vnet.ibm.com