serial_pty_log: fix log decode failure
When booting OVMF guest, there are some un-decoded special characters in the
console log which caused file reading failed. So the solution is to add this
error='ignore' parameter in order to not raise the error like below:
UnicodeDecodeError: 'utf-8' codec can't decode byte 0xd9 in position 0: unexpected end of data

The ignoring will not impact some key strings matched in the log.

Signed-off-by: Dan Zheng <dzheng@redhat.com>