Avoid GC for the diskxml object
The XML file to attach a disk is managed by the object named "diskxml" in
create_disk_xml function. But this function returns only diskxml.xml member,
it is a mere a string. So, the diskxml object lose its reference after
the function returning the value.
Because of this, diskxml object may be destroyed by the python gabage
collection.

If the diskxml object is destroyed, it removes related XML files in
its finalizer. It causes test failure with "No such file or directory"
error when it executes virsh attach-device command.