libvirtd: Ignore PVR err str for aarch64
The PVR err is output during CPU detection for ARM. Which was caused by the
missing cpu xml in libvirt/src/cpu_map.

CPU detection will (in some cases) report the host CPU in host capabilities,
but that's it. It's not very useful so we could ignore the PVR err.
Otherwise we need to add all arm cpu to libvirt/src/cpu_map.

Ref: https://listman.redhat.com/archives/libvir-list/2022-August/234040.html

Signed-off-by: Liu Yiding <liuyd.fnst@fujitsu.com>