virt_sysprep: check ssh v2 public key instead of v1
"ssh_host_key.pub" is a key file for SSHv1 protocol, that is deprecated
in 2017 and not exist in today's distributions.
Instead, "ssh_host_rsa_key.pub" should be used.

This test checks the contents of the key file to confirm virt-sysprep
command propery removes the key file. If the file doesn't exist,
this test cannot confirm the file deletion because the check always
return "No such file or directory" error.