virt_fil_operations: Add a trailing newline when retreive contents
Same fix as PR#4468

The cmd_output function in aexpect module trim the last line because it
includes shell prompt.
https://github.com/avocado-framework/aexpect/blob/master/aexpect/client.py#L1202

In this test, the content of guestfs_temp has only one line and it doesn't
include newline at the end. So, the cat output is followed by the next
prompt in one line and trimmed in remove_last_nonempty_line function.
Because of this, the file content seems to be empty and test fails.

This patch adds additional newline after the cat command so that the
contents of the file is not trimmed by the remove_last_nonempty_line
function.