passthrough/pci: add test for ISM device on s390x
ISM is a special device on s390x that can be passed through
with vfio_pci. It's managed by libvirt as node device.

Add a test case that confirms that the device can be attached
and is listed inside the guest.

Use a placeholder in the configuration file for device selection.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>