v2v: fix firmware check issue
The old firmware check is as a public checkpoint, but in fact, not
all VMs have that attribute. And the original test case and bug only
report the attribute should be show in xml only for vmware guests.

So I moved the check for all test cases and put it into a specific
test case.

Signed-off-by: Xiaodai Wang <xiaodwan@redhat.com>