Fix for virsh_capabilities when tcg is disabled on power
Signed-off-by: Kowshik Jois B S <kowsjois@linux.ibm.com>

Description: virsh URI doesn't discriminate between TCG and KVM, so it contains the string "qemu" in both the cases. Which is correct. But when TCG is disabled, there will not be a domain type = 'qemu'.

The correct way to check this is to parse the output of $QEMU_BIN --accel help and cross check that with what virsh capabilities shows.