libvirt_rng: Ignore non-virtio rng device
The tests compare the number of available rng devices
to confirm attachment. However, some sources can provide
rng devices that are not virtio and can be ignored.

Instead of adding logic to the code that ignores one by one,
change the code to provide a configurable list of devices
that should be ignored.

The criterion to ignore the device 'tpm-rng-0' was its presence
in the XML configuration. However, it should always be ignored.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>