[incremental_backup] Fix failures for clean up checkpoints
For negative casese which intentionally destroy vm or kill qemu
process during backup process, the checkpoints' metadata and qcow2
files' data could be inconsistent. And in this situation when case
tries to cleanup checkpoints without '--metadata', error may
happen. This fix is to delete chcekpoints with '--metadata' in
the cleanup process for these negative cases.

Signed-off-by: Yi Sun <yisun@redhat.com>