Update the tx_queue_size parameter
After qemu bug 2040509, qemu changed the bahavior when checking
the tx_queue_size value. Only for vdpa and vhostuser, the tx_queue_size
more than 256 are supported. For other types, it will report error.
Before this bug, qemu will ignore the unsupported setting and start
the vm anyway and always keep tx_queue_size as 256 in the VM. This
may be supported in the future, refer to
https://issues.redhat.com/browse/RHEL-1139. Update this parameter
setting as it is not critical for these cases.

Signed-off-by: Yalan Zhang <yalzhang@redhat.com>