libvirt_rng: remove flakiness for tcp bind mode
Sometimes,

 libvirt_rng.backend_tcp.bind_mode.hotplug_unplug.positive.persistent

fails.

The test provides entropy to the tcp backend via netcat

 nc -4 localhost 1024

However, this command will immediately exit if nothing listens on that
port.
The test will hotplug an egd device which listens. However, if netcat
is issued before the listener is active then netcat will exit and
that egd device doesn't receive entropy during the boot. Therefore the
boot will wait and the test can't log into the console.

Change execution order to make sure the entropy is provided, that is,
first hotplug the listener, then send data with netcat.

I assume that on environments where the test passes, the bgjob is
started shortly after device is attached, and that this is possible
because it's started asynchronously because it uses the class AsyncJob.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>