Disable serial device hotplug test on arm
Previously, an error message occured while attempting to hotplug or
unplug serial devices on the arm64 architecture. This is due to the
lack of support for these operations on arm64.
To resolve this issue, this commit refactors the configuration file
to exclude hot-plugging and unplugging of serial devices specifically
for arm64. By doing so, we ensure better compatibility and prevent
unknown errors.

Signed-off-by: Aakanksha <aatripat@redhat.com>