libvirt-guests: Improve test logging
The test sometimes fails and there's no log for the guest service at
all. Add checks for start and restart to succeed and if not, log
the service logs with journalctl.

Also, the Tail was called incorrectly leading to (per default unrecorded) TypeError

"can only concatenate str (not "tuple") to str"

The Tail already has a log file where it stores the output. Print its
path so it can be inspected if needed.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>