virtual_disks_multidisks: fix disk discovery
The tests used the disk target as identifier for the
disk inside of the VM. However, the target is not guaranteed
to show up in the guest as-is, especially given that the
/dev/XdY paths are not persistent and can change between boots.

Use instead a function that identifies a new disk inside of a VM
as the one disk that has no root mount on itself, its partitions
or volumes.

Also, the delete_scsi_disk doesn't return any value so it
was always looping for the full timeout. Wrap it into a function
that confirms if both the module and device have been removed.

Finally, disable floppy testing for s390x.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>