numa: Fix the check of guest numa cell allaction
If the memory mode and memnode mode are not set, It is unpredictable on which
host numa node a guest numa cell will be allocated. So all the host numa nodes
should be checked.

The current test only considers the case where the host has two numa nodes.
This patch fixes this issue.

Signed-off-by: Hu Shuai <hus.fnst@fujitsu.com>