nmi_test: disable second inject for s390x
On s390x the inject NMI behavior is different to other archs:
It panics and if kdump is configured the kernel will be dumped and the
guest reboots.

Both will result in the NMI inject count 0.

A second quick injection will result in a panic because kdump is not up
yet.

Skip the qemu-monitor command, only check the libvirt command.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>