blkdeviotune: improve test log
The test case needs the output from 'blkdeviotune' on the disk in order
to compare it with the XML values.

It was trying to compare with an empty dictionary because the result was
empty with an error that was ignored.

Don't continue the test if the block I/O throttle parameters cannot be
retrieved. This will show a test error indicating this.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>