passt: fix check_default_gw
The default gateway function in utils_net now will return
None when it detects ipv6 multipath.

Update the provider code to fail the test if this is the case.
If not, the test case will fail with "NoneType" object has no
attribute 'split'".

Also, pass the host interface to be tested to have more control.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>