emulatorpin: enable test case always
The test case is valuable even if the number of host cpus
is not a multiple of 8 which is a special subcase of issue
reported in https://bugzilla.redhat.com/show_bug.cgi?id=1227180

Instead of skipping if this is not a case run it still.

We'll have an internal test doc to document the behavior
in case of multiple of 8.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>