libvirt_rng: revert commit to avoid timeouts
Before commit 40b1324887851acc3abc1c88bc250faea279b032
dd read and wrote 512 bytes at a time (bs default), 100 times.
This led to a total size of 51200 bytes = 0.512 KB = 0.000512 MB.

The commit set instead 1M = 1MB to be written and read at a time, 10
times. This would result in a total size of 10 MB which seems to be
a typo because 10 MB are 10 000 000 bytes and the set rate limit is
2500 bytes per second in the rng_rate test cases; the test would
most definitely timeout because 'dd' will need at least 4000 seconds,
that is, more than 1 hr, to finish.

Because of this commit we now experience timeouts on VM that performed
well, so revert it.

If there are timeouts with the original command, it is likely because
the backends don't create enough entropy. Making sure that rngd is
running on the test host should solve this for two major backend scenarios
that use /dev/random, that is:

a) model='random', /dev/random
b) model='egd' type='tcp': the test feeds it from /dev/random, s. code

model='builtin' uses a QEMU-internal generator. The rate check tests
the upper limit, so if a generator doesn't provide at a higher rate
any rate is still valid, so in those cases we should test a lower rate
limit.

However, on request, to avoid timeouts on low performance systems
set count to 10. This will create only 5KB and for the rate_limit
cases this would mean with a limit of 2500 B/s ~ 2.4 KB/s
2-3 seconds will be sufficient. On the other hand, as the limiting
can result in higher throughput initially, this amount might not
be enough on some systems to average out the initial deviation.
Therefore, make the throughput configurable in case the rate_limit
check fails on specific systems.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>