libvirt_rng: fix variable 'output' referenced before assignment
When aexpect.exceptions.ShellTimeoutError is raised and the case
is rng_rate, the UnboundLocalError: local variable 'output'
referenced before assignment will be reported.
Because when the timeout happens, the dd rate will not be reported
in dd cmdline, it has no way to calc the rate. So it's reasonable
to fail the test when timeout happens.

Signed-off-by: Xiaodai Wang <xiaodwan@redhat.com>