rerror_policy: fix disk selection
The disk was originally selected by its target name but those names
are impersistent and might change in the guest.
The test would then sometimes fail as it would read (without I/O errors)
from the system disk instead of the new nbd disk when the system was at
vdb.
Instead, discover the new disk as the one that doesn't have the root
filesystem.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>