virsh_snapshot_disk: log vm state
Sometimes tests in CI fail with
"""
error: XML error: memory state cannot be saved with offline or disk-only
snapshot.
"""
but I cannot reproduce it.

I assume the VM is not in the correct state as disk-only snasphot is
not requested in the test case.

Log the VM state before running the snapshot-create command to check it
next time the issue arises.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>