virtio_page_per_vq: select input device on type
The tests statically selected the input element of various
by its expected list index. This failed if the domain xml
was different.

Select now the input element by its type and the expected driver
subelement to avoid relying on order which is not defined by
the XML schema here.

Finally, print more info to log for easier debugging.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>