virsh_set_user_password: fix openssl command
As we've upgraded to openssl 3.2 the condition selects the incorrect
command causing an error "-crypt" flag doesn't exist.

Assume that the API won't change in 3 to select the right command.

While at it, only try to create the test user if it doesn't already
exist, 'id -u <test-user>' returns 1 if it doesn't.

Apply 'black' reformatting to avoid CI issues.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>