requirements-travis: fix dependencies
The systems can't be set up anymore for python versions
since 3.8. Update depdendencies.

There will be a later pylint version producing new errors.
Ignore these as we didn't check them before.

Update pip to latest to avoid other dependency issues.

Allow for pylint to exit with code 32 if there are no files to
check. Otherwise the build will fail.
For that, the exit-on-error must be temporarily switched off (set +e)
in the run commands.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>