vcpu_metrics.py: Set number of queues for interface
- The Problem

libvirt does not support more than one interface queue
for hosts using <interface type='user'>

If the xml that is being used for this test sets the
number of queues higher, then the test will fail.

- Solution

By removing the part of the XML that sets the number
of queues. Libvirt can default to a working number of
queues

Signed-off-by: Slancaster1 <slancast@redhat.com>