security/virt_what: add test for expected value
Since virt-what-1.25-9, there's an additional command 'virt-what-cvm'
that gives information about if and which confidential VM mode is
running.

The test per default assumes that the VM runs without confidentiality.
It can be configured to cover other values depending on the guest.
The supported outputs are listed on the man virt-what-cvm.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>