Virtio_packed: add new test scenarios of driver packed attribute
Automate VIRT-293723 - [packed] Start vm with virtio packed attribute
and check the bit. After discussion, we don't need to add the checkpoint
of bit. Also because this attribute have already been covered in some
devices but not filesystem, video and others. So I covered all them
directly in the previous test script.

Signed-off-by: Meina Li <meili@redhat.com>