boot_from_usb_device.cfg: Update regex
- The Problem
The test guest_os_booting.boot_order.usb_device.block_device
looks for a regex to see if the usb boot device is being booted
first.

If the usb device is blank, edk2 will reorder the boot order,
to deprioritize the usb device. And it won't print it first
in the boot order.

- The Solution

We can see that it attempts to boot from the usb
if we see edk2 printing the "USB_BOOT_TEST_UNIT_READY"
command output.

```
UsbBootExecCmd: Success to Exec 0x0 Cmd (Result = 1)
```

Cmd 0x0: Test Unit Ready Opcode
Result = 1: The unit was ready for further commands

This is printed when edk2 is asking the USB device
if it is ready to recieve further commands.

We can simply test for this output to see if
edk2 considered the usb device for booting.

Signed-off-by: Slancaster1 <slancast@redhat.com>