svirt/selinux/per_device: return if expected error is found
`status_error` expected to be `yes` on invalid configuration, that is,
either during xml setup

a) through `vmxml.sync` which undefines and tries to define the VM, or
b) through attach through `virsh.attach_device`.

Both will raise LibvirtXMLError on failure. So, if as expected this
error is raised for coldplug, we can stop the test execution by returning
and considering the test PASS.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>