Get the proper numa nodes configuration from the system.
The Patch fixes almost 67 cases which were failing as Nn numa nodes available in the system.

The provided fix check for all the available numa nodes in the system and figures out which numa node has memory an cpus.

Even the system is having randome numa nodes configurations the fix works for all the manners and gets the list nodes which has memory and cpus.

Signed-off-by: Tasmiya Nalatwad <tasmiya@linux.vnet.ibm.com>