virsh-snapshot: Update logic to verify snapshot
check_snaplist function is used to verify the created snapshot
by checking it's presence in the guest xml.
But since there might be multiple disk name values, drivers, source
files, snapshot types, etc, the verification is failing leading to
false negative test result.

This patch will allow us to check the disk values by checking if it
is present in the list of xml parameter values.

Signed-off-by: Misbah Anjum N <misanjum@linux.vnet.ibm.com>