iface_network/dhcp_range: expect second guest not to get IP
In 1e88841514 `utils_net.get_guest_ip_addr` was replaced by
`network_base.get_vm_ip`. However, the first function returned
`None` when IP wasn't available whereas the second function
raises a test error. Update the code to handle the test error
in the same way it would have handle a return value `None`.

In this specific case, when dhcp_range is tested, the range is
very small so the second VM is expected not to get an ip.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>