get_vm_ip:Return None if needed
We use get_vm_ip to replace utils_net.get_guest_ip_addr in some test
code to avoid potential risks, but it changes the original behavior
which is to return None when cannot get valid ip address. Add an argument
to choose between returning None or raising error to fix this problem.

Signed-off-by: Haijiao Zhao <haizhao@redhat.com>