sriov: fix ipxe rom file issue
Relying on ipxe build in tests when cannot obtain VF pxe rom file is an unstable testing method. Because this case is mainly to test the xml but not the pxe function. So we decided to use the fixed virtio rom file to test.
The error message:
This test needs rom file: /usr/share/ipxe/8086154c.rom.

Signed-off-by: Meina Li <meili@redhat.com>