transitional_vsock: disable on s390x
The test assumes that vsock is a PCI device with additional device
information to be checked. Only one test variant applies to s390x
(virtio).

We already have test vsock.* that covers this setup and more as it also
covers functionality (establish communication).

Therefore, skip this test suite for s390x.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>