virsh_detach_device_alias: enhancement for pci
Currently the cases are skipped in most time because below error:

error: Failed to attach device from pci.xml
error: internal error: Non-endpoint PCI devices cannot be assigned to guests

This is because the incorrect pci device is used to attach to the vm.
This fix is an enhancement to increase possiblities of finding the suitable pci device.

Signed-off-by: Dan Zheng <dzheng@redhat.com>